Moved file system mock into seperate module
