+++ plugin/__mocks__/fs/promises.ts
@@ -1,0 +1,33 @@
+//Must be before imports, because mock-definitions are hoisted
+async function mockFsRead (path: string): Promise<Buffer> {
+  switch (path) {
+    case '/simple.html': {
+      return Buffer.from('<h1>HelloWorld<h1>')
+    }
+
+    case '/with-partial.html': {
+      return Buffer.from('<p><vite-partial src="/simple.html"/></p>')
+    }
+
+    case '/subdirectory/with-relative-partial.html': {
+      return Buffer.from('<p><vite-partial src="./../simple.html"/></p>')
+    }
+
+    case '/subdirectory/with-absolute-partial.html': {
+      return Buffer.from('<p><vite-partial src="/simple.html"/></p>')
+    }
+
+    case '/root/index.html': {
+      return Buffer.from('<p><vite-partial src="/partial.html"/></p>')
+    }
+
+    case '/root/partial.html': {
+      return Buffer.from('<h1>HelloWorld<h1>')
+    }
+
+    default:
+      throw new Error('File not found')
+  }
+}
+
+export default { readFile: mockFsRead }
