+++ plugin/setupTests.ts
@@ -1,0 +1,6 @@
+import { vi } from 'vitest'
+import fsPromisesMock from './__mocks__/fs/promises'
+
+vi.mock('fs/promises', async () => {
+  return fsPromisesMock;
+})
