+++ plugin/vite.config.ts
@@ -1,0 +1,11 @@
+import { defineConfig } from 'vite';
+
+export default defineConfig(() => {
+  return {
+    test: {
+      environment: 'node',
+      globals: true,
+      setupFiles: ['./setupTests.ts'],
+    },
+  };
+});
